[Ionut] - Linux port work
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@454 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 2813e1c5eefbe2f66797ef1cbb4f7e43156136f5 [formerly 90063fc1ae5bd61b2e60800209bc290230ca5f50] [formerly 31954e333678403f9d833d9098194e598f90fbb7 [formerly ca88af727bebc9a152baa613ce1ddae80f0f6798]]
Former-commit-id: b674c4cc64b2ed2bd176528bfc4be6f4700b5373 [formerly 846722e0c207fc712012025dbae3642f61403999]
Former-commit-id: d91809ac5b91b9f8209f21e366c28c6320c14957

+++ Source Code/Core/Math/BoundingVolumes/Headers/BoundingBox.inl
@@ -203,7 +203,7 @@ inline const vec3<F32>* BoundingBox::getPoints() const {
 }
 
 inline F32 BoundingBox::nearestDistanceFromPoint(const vec3<F32>& pos) const {
-    return std::sqrtf(nearestDistanceFromPointSquared(pos));
+    return std::sqrt(nearestDistanceFromPointSquared(pos));
 }
 
 inline void BoundingBox::ComputePoints() const {
